+++ sh/fs.stats.sh
@@ -1,5 +1,0 @@
-#!/bin/bash
-cd /var/www/production/current/
-FS_ENV=prod php-production run.php Stats bezirke
-FS_ENV=prod php-production run.php Stats betriebe
-FS_ENV=prod php-production run.php Stats foodsaver